ingress/requirements.txt
pika==0.12.0
websocket-client==0.53.0